scripts-testing/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.36.1
pandas==2.0.0
matplotlib==3.7.1
tabulate==0.9.0
